requirements.txt
alembic==1.9.2
anyio==3.6.2
attrs==22.2.0
bidict==0.22.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
dnspython==2.3.0
ecdsa==0.18.0
eventlet==0.33.3
exceptiongroup==1.1.0
fastapi==0.89.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
jose==1.0.0
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
passlib==1.7.4
Pillow==9.4.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
pytest==7.2.1
python-engineio==4.3.4
python-jose==3.3.0
python-multipart==0.0.5
python-socketio==5.7.2
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
